+++ packages/adminui-pub/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - UI Tag tracking for all forms and ExtJS tables what mounts as tab:
+   - `$App.doCommand` generate a UI Tag for and assign it to uiTag property of central panel tabs
+   - on tab activation `beforetabchange` event handler sets a `connection.setUiTag(tab.uiTag)`
+   - `uitag=${uiTag}` is added to all `ubql` requeset URL, so in server logs the source of UBQL can be tracked   
 
 ### Changed